+++ src/icu/normalizer/mod.rs
@@ -1,50 +1,13 @@
-use std::mem;
-
 use rust_icu_unorm2::UNormalizer;
-use tantivy::tokenizer::{BoxTokenStream, Token, TokenFilter, TokenStream};
 
-struct ICUNormalizer2TokenStream<'a> {
-    normalizer: UNormalizer,
-    tail: BoxTokenStream<'a>,
-    temp: String,
-}
+use super::Error;
+pub use token_filter::ICUNormalizer2TokenFilter;
+use token_stream::ICUNormalizer2TokenStream;
+use wrapper::ICUNormalizer2FilterWrapper;
 
-impl<'a> TokenStream for ICUNormalizer2TokenStream<'a> {
-    fn advance(&mut self) -> bool {
-        let result = self.tail.advance();
-        if !result {
-            return false;
-        }
-
-        if let Ok(t) = self.normalizer.normalize(&self.tail.token().text) {
-            self.temp = t;
-            mem::swap(&mut self.tail.token_mut().text, &mut self.temp);
-        }
-        result
-    }
-
-    fn token(&self) -> &Token {
-        self.tail.token()
-    }
-
-    fn token_mut(&mut self) -> &mut Token {
-        self.tail.token_mut()
-    }
-}
-
-impl From<Mode> for UNormalizer {
-    fn from(tp: Mode) -> Self {
-        match tp {
-            Mode::NFC => UNormalizer::new_nfc().expect("Can't create NFC normalizer"),
-            Mode::NFD => UNormalizer::new_nfd().expect("Can't create NFD normalizer"),
-            Mode::NFKC => UNormalizer::new_nfkc().expect("Can't create NFKC normalizer"),
-            Mode::NFKD => UNormalizer::new_nfkd().expect("Can't create NFKD normalizer"),
-            Mode::NFKCCasefold => {
-                UNormalizer::new_nfkc_casefold().expect("Can't create NFKC casefold normalizer")
-            }
-        }
-    }
-}
+mod token_filter;
+mod token_stream;
+mod wrapper;
 
 /// Normalization algorithms (see [Wikipedia](https://en.wikipedia.org/wiki/Unicode_equivalence#Normalization)).
 #[derive(Clone, Debug, Copy)]
@@ -61,61 +24,18 @@ pub enum Mode {
     NFKCCasefold,
 }
 
-/// [TokenFilter] that converts text into a normal form.
-/// It supports all [Google's unicode normalization](https://docs.rs/rust_icu_unorm2/2.0.0/rust_icu_unorm2/struct.UNormalizer.html) using [Mode]:
-/// * NFC
-/// * NFD
-/// * NFKC
-/// * NFKD
-/// * NFKC casefold
-///
-/// See Wikipedia's [unicode normalization](https://en.wikipedia.org/wiki/Unicode_equivalence#Normalization) or
-/// [Unicode documentation](https://www.unicode.org/reports/tr15/) for more information.
-///
-/// Building an [ICUNormalizer2TokenFilter] is straightforward :
-/// ```rust
-/// use tantivy_analysis_contrib::icu::ICUNormalizer2TokenFilter;
-/// use tantivy_analysis_contrib::icu::Mode;
-///
-/// let normalizer = ICUNormalizer2TokenFilter {
-///     mode: Mode::NFD,
-/// };
-/// ```
-///
-/// # Example
-///
-/// Here is an example showing which tokens are produce
-///
-/// ```rust
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// use tantivy::tokenizer::{RawTokenizer, TextAnalyzer, Token};
-/// use tantivy_analysis_contrib::icu::{ICUNormalizer2TokenFilter, Mode};
-///
-/// let mut token_stream = TextAnalyzer::from(RawTokenizer)
-///             .filter(ICUNormalizer2TokenFilter { mode: Mode::NFKCCasefold })
-///             .token_stream("Ruß");
-///
-/// let token = token_stream.next().expect("A token should be present.");
-///
-/// assert_eq!(token.text, "russ".to_string());
-///
-/// assert_eq!(None, token_stream.next());
-/// #     Ok(())
-/// # }
-/// ```
-#[derive(Clone, Copy, Debug)]
-pub struct ICUNormalizer2TokenFilter {
-    /// Normalization algorithm.
-    pub mode: Mode,
-}
+impl TryFrom<Mode> for UNormalizer {
+    type Error = Error;
 
-impl TokenFilter for ICUNormalizer2TokenFilter {
-    fn transform<'a>(&self, token_stream: BoxTokenStream<'a>) -> BoxTokenStream<'a> {
-        From::from(ICUNormalizer2TokenStream {
-            normalizer: UNormalizer::from(self.mode),
-            tail: token_stream,
-            temp: String::with_capacity(100),
-        })
+    fn try_from(tp: Mode) -> Result<Self, Self::Error> {
+        let normalizer = match tp {
+            Mode::NFC => UNormalizer::new_nfc()?,
+            Mode::NFD => UNormalizer::new_nfd()?,
+            Mode::NFKC => UNormalizer::new_nfkc()?,
+            Mode::NFKD => UNormalizer::new_nfkd()?,
+            Mode::NFKCCasefold => UNormalizer::new_nfkc_casefold()?,
+        };
+        Ok(normalizer)
     }
 }
 
@@ -123,14 +43,17 @@ impl TokenFilter for ICUNormalizer2TokenFilter {
 mod tests {
     use std::error::Error;
 
-    use tantivy::tokenizer::{RawTokenizer, TextAnalyzer, WhitespaceTokenizer};
+    use tantivy::tokenizer::{RawTokenizer, TextAnalyzer, Token, WhitespaceTokenizer};
 
     use super::*;
 
     fn token_stream_helper(text: &str, tp: Mode) -> Vec<Token> {
-        let mut token_stream = TextAnalyzer::from(WhitespaceTokenizer)
-            .filter(ICUNormalizer2TokenFilter { mode: tp })
-            .token_stream(text);
+        let mut a = TextAnalyzer::builder(WhitespaceTokenizer::default())
+            .filter(ICUNormalizer2TokenFilter::new(tp).unwrap())
+            .build();
+
+        let mut token_stream = a.token_stream(text);
+
         let mut tokens = vec![];
         let mut add_token = |token: &Token| {
             tokens.push(token.clone());
@@ -140,9 +63,12 @@ mod tests {
     }
 
     fn token_stream_helper_raw(text: &str, tp: Mode) -> Vec<Token> {
-        let mut token_stream = TextAnalyzer::from(RawTokenizer)
-            .filter(ICUNormalizer2TokenFilter { mode: tp })
-            .token_stream(text);
+        let mut a = TextAnalyzer::builder(RawTokenizer::default())
+            .filter(ICUNormalizer2TokenFilter::new(tp).unwrap())
+            .build();
+
+        let mut token_stream = a.token_stream(text);
+
         let mut tokens = vec![];
         let mut add_token = |token: &Token| {
             tokens.push(token.clone());